Move `ModelType` out of api.
