+++ executor/src/test/java/org/creekservice/internal/system/test/executor/api/test/model/TestModelTest.java
@@ -40,7 +40,7 @@ import org.creekservice.api.system.test.extension.test.model.ExpectationRef;
 import org.creekservice.api.system.test.extension.test.model.Input;
 import org.creekservice.api.system.test.extension.test.model.InputHandler;
 import org.creekservice.api.system.test.extension.test.model.InputRef;
-import org.creekservice.api.system.test.extension.test.model.ModelType;
+import org.creekservice.api.system.test.parser.ModelType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
