+++ parser/src/main/java/org/creekservice/internal/system/test/parser/SystemTestMapper.java
@@ -38,9 +38,9 @@ import org.creekservice.api.system.test.extension.test.model.Expectation;
 import org.creekservice.api.system.test.extension.test.model.ExpectationRef;
 import org.creekservice.api.system.test.extension.test.model.Input;
 import org.creekservice.api.system.test.extension.test.model.InputRef;
-import org.creekservice.api.system.test.extension.test.model.ModelType;
 import org.creekservice.api.system.test.extension.test.model.Ref;
 import org.creekservice.api.system.test.model.LocationAware;
+import org.creekservice.api.system.test.parser.ModelType;
 
 public final class SystemTestMapper {
