+++ parser/src/test/java/org/creekservice/internal/system/test/parser/SystemTestMapperTest.java
@@ -41,9 +41,9 @@ import org.creekservice.api.system.test.extension.test.model.Expectation;
 import org.creekservice.api.system.test.extension.test.model.ExpectationRef;
 import org.creekservice.api.system.test.extension.test.model.Input;
 import org.creekservice.api.system.test.extension.test.model.InputRef;
-import org.creekservice.api.system.test.extension.test.model.ModelType;
 import org.creekservice.api.system.test.extension.test.model.Ref;
 import org.creekservice.api.system.test.model.LocationAware;
+import org.creekservice.api.system.test.parser.ModelType;
 import org.creekservice.api.test.util.TestPaths;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
